+++ game-devilmaycry5/CHANGELOG.md
@@ -2,8 +2,7 @@
 
 ## Planned Improvements (Not Yet Released)
 
-- Fixed: Mod files with no file extension were skipped during installation
-- Added an "Open SteamDB Page" button next to the PCGamingWiki one.
+- None
 
 ## [1.0.0] - 2026-08-12